psychobot/requirements.txt
aiogram==3.3.0
python-dotenv==1.0.0
openai==1.12.0
sqlalchemy==2.0.25
asyncpg==0.29.0
aiohttp==3.9.3
python-dateutil==2.8.2 
